requirements.txt
mysql-connector-python==8.0.29
pandas==1.3.4
matplotlib=3.4.3